requirements_fixed.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
openai==1.3.7
python-multipart==0.0.6
pandas>=2.2.0
pydantic==2.5.0
python-dotenv==1.0.0
codecarbon==2.2.4
httpx==0.25.2
numpy>=1.26.0
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
